requirements.txt
altgraph==0.17.3
astroid==2.12.12
autopep8==2.0.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
decorator==4.4.2
dill==0.3.6
idna==3.4
imageio==2.22.4
imageio-ffmpeg==0.4.7
instagrapi==1.16.30
isort==5.10.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
moviepy==1.0.3
numpy==1.23.4
Pillow==9.3.0
platformdirs==2.5.3
proglog==0.1.10
pycodestyle==2.9.1
pycparser==2.21
pycryptodomex==3.15.0
pydantic==1.10.2
pyinstaller-hooks-contrib==2022.13
pylint==2.15.5
PySimpleGUI==4.60.4
PySocks==1.7.1
python-dotenv==0.21.0
requests==2.28.1
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
wrapt==1.14.1
